+++ packages/core/src/components/DateRangePicker/helpers/index.ts
@@ -1,1 +1,2 @@
+export * from './dateRangeHelpers';
 export * from './getFormattedDate';