+++ packages/core/src/components/HelperAndErrorTextTooltip/types.ts
@@ -1,0 +1,8 @@
+export interface HelperTextTooltipProps {
+    /** Id prefix */
+    id: string;
+    /** Helper Text */
+    helperText?: string;
+    /** Error Text */
+    errorText?: string;
+}