+++ packages/layout/src/components/Header/Header.context.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+
+export interface Context {
+    isOpen: boolean;
+    onClose: () => void;
+}
+
+const HeaderContext = React.createContext<Context>(null);
+
+export default HeaderContext;