+++ packages/core/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.16.0](https://github.com/medly/medly-components/compare/@medly-components/core@2.15.1...@medly-components/core@2.16.0) (2021-08-17)
+
+
+### Features
+
+* **core:** add custom range selection options in date range picker ([#382](https://github.com/medly/medly-components/issues/382)) ([aece055](https://github.com/medly/medly-components/commit/aece0558151c1c194152e34cd0103887bf4c33e6))
+
+
+
+
+
+## [2.15.1](https://github.com/medly/medly-components/compare/@medly-components/core@2.15.0...@medly-components/core@2.15.1) (2021-08-17)
+
+**Note:** Version bump only for package @medly-components/core
+
+
+
+
+
+# [2.15.0](https://github.com/medly/medly-components/compare/@medly-components/core@2.14.0...@medly-components/core@2.15.0) (2021-08-16)
+
+
+### Features
+
+* **layout:** new header layout component ([#377](https://github.com/medly/medly-components/issues/377)) ([d38bd12](https://github.com/medly/medly-components/commit/d38bd12b5f903e9ee29d61b812afc8c17dc617d3))
+
+
+
+
+
+# [2.14.0](https://github.com/medly/medly-components/compare/@medly-components/core@2.13.1...@medly-components/core@2.14.0) (2021-08-13)
+
+
+### Bug Fixes
+
+* **core:** fix imports ([017f49e](https://github.com/medly/medly-components/commit/017f49e15708c4903134eb418725015656b18c48))
+
+
+### Features
+
+* small updates for compatibility with character count ([3da12ac](https://github.com/medly/medly-components/commit/3da12ac806cfb60296d8160b970085ed258b1c72))
+* **core:** add a new variant for helper/error text ([7cf0451](https://github.com/medly/medly-components/commit/7cf045139cfab42510fa6a270bb19eaf329f2bae))
+* **core:** create helperAndErrorTextTooltip component and use it in all fields ([c58f598](https://github.com/medly/medly-components/commit/c58f598bd0c76ba3f35ccf94593f5b51d97ee7a3))
+
+
+
+
+
 ## [2.13.1](https://github.com/medly/medly-components/compare/@medly-components/core@2.13.0...@medly-components/core@2.13.1) (2021-08-12)